Tdd: Se modifica función de compresion para que test pase

+++ srcs/Logica/Dibujo.py
@@ -13,7 +13,7 @@ class Pintable(ABC):
 class Dibujo(Pintable):
     
     def __init__(self , x ,y):
-        self.boceto = numpy.zeros((x, y))
+        self.boceto = numpy.zeros((x, y),dtype=int)
 
     def pintar(self, x, y,color):
         self.boceto[x][y] = color
